main: adjust for a focus problem.
-When in fullscreen, and using the file open dialog to select
a new video; if you select the video after the 5 second time out
of the onscreen controls (they have disapeared ), mouse input focus
seems to be lost on the entire window. I suspect that it is supposed
to return to the "file open" button in the toolbar but since it is
hidden, it can not.

-This fix prevents the controls from disapearing while any dialog
is open. The 5 second timeout is retriggered after the dialog
is closed.