Integrate building for Raspbian.
-Test with 8-16-2017 release of Raspbian Stretch.
-Raspbian build is GTK2
-Ubuntu Mate build is GTK3

-Due differences in the window manager? in Raspbian, I'm having
to poll for the window position and calculate the render_window
position manually. In Ubuntu Mate this is not needed because
we get window position events while dragging the window. I
use the define POLLWINPOS to switch the code for the case.

-Include the original GIMP .xcf for the icon. Maybe someone
will want to theme it.

-Include a blue icon for Raspbian.

-Include "Installed-Size" in Debian control file.

-Register signal handlers in Ubuntu Mate to make sure we
kill omxplayer if we are going to crash... Does not work
properly and may not be needed in Raspbian. The define
USE_SIGHANDLER is used to switch the code.

-Various other small fix ups.