Render window consistency after transition.
-When transitioning to a new video, we want to be consistant
with the state of the render window being hidden or transparent.
For example, if you have 'Continuous Playback' enabled and also
have a dialog window showing; the render window is hidden and if
the video ends and a new one begins, we don't want the
render window to then, show over the dialog window.