op_control: fix omxplayer doesn't always exit.
-I found a case in which, during a playlist loop, if you
drag the slider to the end; then wait for the next video to
start; then drag the slider to the end again. omxplayer would
not exit properly, leaving everthing in sort of a defunct state.

-So, with this new code we are, creating a new thread to read
the output of omxplayer, and checking for the 'have a nice day'
which is printed at the end.

-During 'opc_stop_omxplayer()', we only want the playback_completed
cb to fire if the playback ended without user intervention. So,
we temporarily set the call back null while we stop playback
threads.