top_widget: account for high latency and more uri schemes.
-I was not aware that omxplayer could stream video from so
many different uri schemes. So, I only accounted for http and
https. Looking at OMXReader.cpp in the omxplayer project shows
that all of these protcols/uri schemes are supported.

-Remove timeout when launching omxplayer. Now, restore_volume will
loop forever unless a new file path or url is started, in which
case the thread for restore_volume will be restarted.